Cleanup <App />
