+++ js/components/App/Wrapper.js
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+import constants from '../../constants';
+
+export default styled.section`
+  max-width: ${constants.maxWidth};
+  margin: 0 auto;
+  display: block;
+  border-radius: 5px;
+  background-color: white;
+  margin-bottom: 2em;
+  box-shadow: 0 2px 2px 0 rgba(0,0,0,.14),0 3px 1px -2px rgba(0,0,0,.2),0 1px 5px 0 rgba(0,0,0,.12);
+  &:after {
+    content: "";
+    display: block;
+    height: 0;
+    clear: both;
+  }
+`;
